Monobehavior info as well as Quaternion research
Monobehavior is for scripts that will be attached to gameobjects, which means scriptable object scripts are good for scripts which will NOT specifically attach to game objects, as there are unneeded dependencies. Quaternion.Euler(x,y,z) is complicated, it is converting your input into quaternion values which determine the rotations of the repective object on the angles listed, it is converting your values into quaternions behind the scenes for the computer to read.