Most changes are just moving, end of Day-4
using [CustomEditor(typeof(TypeofObjectYouWantACustomEditorFor))] allows you to create custom elements in the inspector of the designated object, in this project I used it to create a button which just calls a method in a script. Another fact is List.AddRange(List/Array), this will add the input list/array elements to the end of the designated List, essentially allowing you to just merge to seperate lists (of the same Object type).